Combine address query mapping functions and tidy up unused mapping functions.  The fix bug in RecordAudit now that field names for addresses have all been updated with a section name and event location use case so that we acn have unique certificate handlebars
