+++ packages/client/src/forms/register/mappings/event-specific-fields/birth/staticLabel.ts
