+++ packages/client/src/forms/register/mappings/event-specific-fields/death/mutation/deceased-mappings.test.ts
@@ -9,7 +9,7 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-import { setDeathRegistrationSectionTransformer } from '@client/forms/register/mappings/fields/death/mutation/deceased-mappings'
+import { setDeathRegistrationSectionTransformer } from '@client/forms/register/mappings/event-specific-fields/death/mutation/deceased-mappings'
 import {
   mockDeathDeclarationData,
   mockDeathRegistrationSectionData
