+++ packages/client/src/forms/register/mappings/event-specific-fields/death/mutation/deceased-mappings.ts
@@ -1,0 +1,50 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+import {
+  IFormField,
+  IFormData,
+  TransformedData,
+  IFormFieldMutationMapFunction
+} from '@client/forms'
+import { cloneDeep } from 'lodash'
+
+export const fieldToDeceasedDateTransformation =
+  (
+    alternativeSectionId?: string,
+    nestedTransformer?: IFormFieldMutationMapFunction
+  ) =>
+  (
+    transformedData: TransformedData,
+    draftData: IFormData,
+    sectionId: string,
+    field: IFormField
+  ) => {
+    if (!draftData[sectionId] || !draftData[sectionId][field.name]) {
+      return transformedData
+    }
+
+    let fieldValue = draftData[sectionId][field.name]
+
+    if (nestedTransformer) {
+      const clonedTransformedData = cloneDeep(transformedData)
+      nestedTransformer(clonedTransformedData, draftData, sectionId, field)
+      fieldValue = clonedTransformedData[sectionId][field.name]
+    }
+
+    transformedData[
+      alternativeSectionId ? alternativeSectionId : sectionId
+    ].deceased = {
+      deceased: true,
+      deathDate: fieldValue
+    }
+    return transformedData
+  }
