--- packages/client/src/forms/register/mappings/fields/event/mutation/documents-mappings.ts
@@ -1,28 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * OpenCRVS is also distributed under the terms of the Civil Registration
- * & Healthcare Disclaimer located at http://opencrvs.org/license.
- *
- * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
- * graphic logo are (registered/a) trademark(s) of Plan International.
- */
-import { fieldToAttachmentTransformer } from '@client/forms/register/mappings/mutation/field-mappings'
-import { TransformedData, IFormData, IFormField } from '@client/forms'
-
-export function eventFieldToAttachmentTransformer(
-  transformedData: TransformedData,
-  draftData: IFormData,
-  sectionId: string,
-  field: IFormField
-) {
-  return fieldToAttachmentTransformer(
-    transformedData,
-    draftData,
-    sectionId,
-    field,
-    'registration'
-  )
-}
