--- packages/client/src/forms/register/mappings/fields/event/query/documents-mappings.ts
@@ -1,29 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * OpenCRVS is also distributed under the terms of the Civil Registration
- * & Healthcare Disclaimer located at http://opencrvs.org/license.
- *
- * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
- * graphic logo are (registered/a) trademark(s) of Plan International.
- */
-import { IFormField, IFormData } from '@client/forms'
-import { attachmentToFieldTransformer } from '@client/forms/register/mappings/query/field-mappings'
-import { EventRegistration } from '@client/utils/gateway'
-
-export function eventAttachmentToFieldTransformer(
-  transformedData: IFormData,
-  queryData: EventRegistration,
-  sectionId: keyof EventRegistration,
-  field: IFormField
-) {
-  return attachmentToFieldTransformer(
-    transformedData,
-    queryData,
-    sectionId,
-    field,
-    'registration'
-  )
-}
