+++ packages/client/src/forms/register/mappings/query/index.ts
@@ -10,14 +10,12 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 export * from './field-mappings'
-export * from '@client/forms/register/mappings/fields/birth/query/father-mappings'
-export * from '@client/forms/register/mappings/fields/birth/query/mother-mappings'
-export * from '@client/forms/register/mappings/fields/birth/query/registration-mappings'
-export * from '@client/forms/register/mappings/fields/death/query/declaration-mappings'
-export * from '@client/forms/register/mappings/fields/death/query/deceased-mappings'
-export * from '@client/forms/register/mappings/fields/death/query/spouse-mappings'
-export * from '@client/forms/register/mappings/fields/marriage/query/registration-mappings'
-export * from '@client/forms/register/mappings/fields/event/query/documents-mappings'
-export * from '@client/forms/register/mappings/fields/informant/query/informant-mappings'
-export * from '@client/forms/register/mappings/fields/marriage/query/event-mappings'
-export * from '@client/forms/register/mappings/fields/customFields/query/customField-mappings'
+export * from './registration-mappings'
+export * from './informant-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/birth/query/father-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/birth/query/mother-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/birth/query/registration-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/death/query/deceased-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/death/query/registration-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/marriage/query/event-mappings'
+export * from '@client/forms/register/mappings/event-specific-fields/marriage/query/registration-mappings'
