+++ packages/client/src/forms/register/mappings/query/utils.ts
@@ -1,0 +1,63 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * OpenCRVS is also distributed under the terms of the Civil Registration
+ * & Healthcare Disclaimer located at http://opencrvs.org/license.
+ *
+ * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
+ * graphic logo are (registered/a) trademark(s) of Plan International.
+ */
+import { GQLRegWorkflow } from '@opencrvs/gateway/src/graphql/schema'
+
+export function transformStatusData(
+  transformedData: IFormData,
+  statusData: GQLRegWorkflow[],
+  sectionId: string
+) {
+  const registrationStatus =
+    statusData &&
+    statusData.find((status) => {
+      return status.type && status.type === 'REGISTERED'
+    })
+  transformedData[sectionId] = {
+    ...transformedData[sectionId],
+    commentsOrNotes:
+      (statusData &&
+        statusData[0] &&
+        statusData[0].comments &&
+        statusData[0].comments[0] &&
+        statusData[0].comments[0].comment) ||
+      ''
+  }
+
+  if (!registrationStatus) {
+    return transformedData
+  }
+  transformedData[sectionId] = {
+    ...transformedData[sectionId],
+    regStatus: {
+      type: registrationStatus.type || '',
+      statusDate: registrationStatus.timestamp,
+      officeName:
+        (registrationStatus.office && registrationStatus.office.name) || '',
+      officeAlias:
+        (registrationStatus.office &&
+          registrationStatus.office.alias &&
+          registrationStatus.office.alias.join(' ')) ||
+        '',
+      officeAddressLevel3:
+        (registrationStatus.office &&
+          registrationStatus.office.address &&
+          registrationStatus.office.address.district) ||
+        '',
+      officeAddressLevel4:
+        (registrationStatus.office &&
+          registrationStatus.office.address &&
+          registrationStatus.office.address.state) ||
+        ''
+    }
+  }
+  return transformedData
+}
